refactor: extract convertArguments into ArgumentConverter utility
